Refactor to divide bridge ERC20 hardhat task into 2 tasks
